recommendations/requirements.txt
grpcio>=1.40.0
grpcio-tools>=1.40.0
pytest # Для будущих тестов